--- a/.circleci/unittest/linux/scripts/run_test.sh
+++ b/.circleci/unittest/linux/scripts/run_test.sh
@@ -7,4 +7,4 @@ conda activate ./env
 
 export PYTORCH_TEST_WITH_SLOW='1'
 python -m torch.utils.collect_env
-pytest --junitxml=test-results/junit.xml -v --durations 2000 -k "test_classification_model or test_detection_model or test_quantized_classification_model or test_segmentation_model or test_video_model"
+pytest --junitxml=test-results/junit.xml -v --durations 20