… test/expect/ModelTester.test_alexnet_expect.pkl
… test/expect/ModelTester.test_convnext_base_expect.pkl
… test/expect/ModelTester.test_convnext_large_expect.pkl
… test/expect/ModelTester.test_efficientnet_b6_expect.pkl
… test/expect/ModelTester.test_efficientnet_b7_expect.pkl
… test/expect/ModelTester.test_efficientnet_v2_l_expect.pkl
… test/expect/ModelTester.test_efficientnet_v2_m_expect.pkl
… test/expect/ModelTester.test_regnet_x_16gf_expect.pkl
… test/expect/ModelTester.test_regnet_x_32gf_expect.pkl
… test/expect/ModelTester.test_regnet_y_128gf_expect.pkl
… test/expect/ModelTester.test_regnet_y_16gf_expect.pkl
… test/expect/ModelTester.test_regnet_y_32gf_expect.pkl
… test/expect/ModelTester.test_resnext101_32x8d_expect.pkl
… test/expect/ModelTester.test_wide_resnet101_2_expect.pkl